refactor(config): Refactor config component
