+++ src/config/impl/index.ts
@@ -1,0 +1,1 @@
+export { buildInstanceOptions } from "./instance-options-impl";
